# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  server.BindUserToRoleRequest:
    properties:
      clusterId:
        type: integer
      roleName:
        type: string
      roleNamespace:
        type: string
      roleType:
        description: ClusterRole 或 Role
        type: string
      userId:
        type: integer
    required:
    - clusterId
    - roleName
    - roleType
    - userId
    type: object
...
  server.UnbindUserFromRoleRequest:
    properties:
      clusterId:
        type: integer
      roleName:
        type: string
      roleNamespace:
        type: string
      userId:
        type: integer
    required:
    - clusterId
    - roleName
    - userId
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 运维管理平台 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: OpsHub API
  version: "1.0"
paths:
...
  /plugins/kubernetes/role-bindings/available-users:
    get:
      consumes:
      - application/json
      description: 获取系统中可以绑定的用户列表
      parameters:
      - description: 搜索关键词
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
      summary: 获取可绑定的用户列表
      tags:
      - Kubernetes/RoleBinding
  /plugins/kubernetes/role-bindings/bind:
    post:
      consumes:
      - application/json
      description: 将平台用户绑定到指定的K8s角色
      parameters:
      - description: 绑定信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server.BindUserToRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
      summary: 绑定用户到K8s角色
      tags:
      - Kubernetes/RoleBinding
  /plugins/kubernetes/role-bindings/credential-users:
    get:
      consumes:
      - application/json
      description: 获取当前用户在指定集群的凭据用户（ServiceAccount）
      parameters:
      - description: 集群ID
        in: query
        name: clusterId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
      summary: 获取集群的凭据用户列表
      tags:
      - Kubernetes/RoleBinding
  /plugins/kubernetes/role-bindings/unbind:
    delete:
      consumes:
      - application/json
      description: 解除用户与K8s角色的绑定
      parameters:
      - description: 解绑信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server.UnbindUserFromRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
      summary: 解绑用户K8s角色
      tags:
      - Kubernetes/RoleBinding
  /plugins/kubernetes/role-bindings/user-bindings:
    get:
      consumes:
      - application/json
      description: 获取指定集群中所有用户（或指定用户）的K8s角色绑定列表
      parameters:
      - description: 集群ID
        in: query
        name: clusterId
        required: true
        type: integer
      - description: 用户ID（可选，不传则返回所有用户）
        in: query
        name: userId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
      summary: 获取用户的所有K8s角色绑定
      tags:
      - Kubernetes/RoleBinding
  /plugins/kubernetes/role-bindings/user-roles:
    get:
      consumes:
      - application/json
      description: 获取指定用户在指定集群的所有角色
      parameters:
      - description: 集群ID
        in: query
        name: clusterId
        required: true
        type: integer
      - description: 用户ID
        in: query
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
      summary: 获取用户在指定集群的角色列表
      tags:
      - Kubernetes/RoleBinding
  /plugins/kubernetes/role-bindings/users:
    get:
      consumes:
      - application/json
      description: 获取指定角色的所有绑定用户
      parameters:
      - description: 集群ID
        in: query
        name: clusterId
        required: true
        type: integer
      - description: 角色名称
        in: query
        name: roleName
        required: true
        type: string
      - description: 角色命名空间
        in: query
        name: roleNamespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
      summary: 获取角色已绑定的用户列表
      tags:
      - Kubernetes/RoleBinding
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"